requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
lets-plot==4.5.1
numpy==2.0.2
palettable==3.3.3
pandas==2.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
